# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User's Rank API",
        "description": "Demo User's Rank based on the Number of Posts API",
        "version": "1.0"
    },
    "paths": {
        "/api/post": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "All User's Posts",
                "description": "All User's Posts",
                "operationId": "f5fce8229f9ae9db81339c7bad1dadb5",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/post/add": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Add Post",
                "description": "Return Post's Insert Status value",
                "operationId": "addNewPost",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title Post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Tag for post, and each tag seperated by ';'",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "Post Content",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Opearation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/post/{postId}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Find Post by ID",
                "description": "Returns a single Post",
                "operationId": "getPostById",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID Post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/post/{postId}/edited": {
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Update Post based on id",
                "description": "Returns a Post's Update Status",
                "operationId": "updatePostById",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID Post",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title Post",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Tag for post, and each tag seperated by ';'",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "Post Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Opearation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/post/{postId}/delete": {
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete Post based on id",
                "description": "Returns a Post's Delete Status",
                "operationId": "deletePostById",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID Post",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Opearation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "description": "Enter token in format (Bearer 'token')",
                "in": "header"
            }
        }
    },
...
}